+++ udata-tabular-preview/explore.vue
@@ -12,7 +12,23 @@
         <caption class="fr-sr-only">{{ $t('Preview of {name}', { name: resource.title }) }}</caption>
         <thead>
           <tr>
-            <th scope="col" v-for="col in columns">{{ col }}</th>
+            <th scope="col" v-for="col in columns">
+              <div class="fr-grid-row fr-grid-row--middle no-wrap">
+                <button
+                  class="fr-btn fr-btn--sm fr-btn--tertiary-no-outline fr-my-n1w"
+                  :class="{
+                    'fr-btn--secondary-grey-500': !isSortedBy(col),
+                    'fr-btn--icon-right': isSortedBy(col),
+                    'fr-icon-arrow-down-line': isSortedBy(col) && sortDesc,
+                    'fr-icon-arrow-up-line': isSortedBy(col) && !sortDesc
+                    }"
+                  @click="sortbyfield(col)"
+                >
+                  {{ col }}
+                  <span class="fr-sr-only">{{ sortDesc ? $t("Sort ascending") : $t("Sort descending") }}</span>
+                </button>
+              </div>
+            </th>
           </tr>
         </thead>
         <tbody>
@@ -42,7 +58,7 @@
 </template>
 
 <script>
-import { apify, changePage, configure, getData } from "@etalab/explore.data.gouv.fr/lib/csvapi";
+import { apify, changePage, configure, getData, sort } from "@etalab/explore.data.gouv.fr/lib/csvapi";
 import { Pagination } from "@etalab/udata-front-plugins-helper";
 import { defineComponent, ref } from 'vue';
 import { tabular_csvapi_url, tabular_page_size } from "./config";
@@ -70,11 +86,31 @@ export default defineComponent({
     const currentPage = ref(1);
     const pageSize = tabular_page_size;
     const totalRows = ref(0);
+    /** @type {import("vue").Ref<string | null>} */
+    const sortBy = ref(null);
+    const sortDesc = ref(false);
+
+    /**
+     *
+     * @param {import("@etalab/explore.data.gouv.fr/lib/csvapi").CsvapiResponse} res
+     */
+    const update = (res) => {
+      if (res.ok) {
+          rows.value = res.rows;
+          columns.value = res.columns;
+          rowCount.value = res.total;
+          columnCount.value = res.columns.length;
+          totalRows.value = res.total;
+          configure({totalRows: totalRows.value});
+        } else {
+          hasError.value = true;
+        }
+    }
 
     const changeExplorePage = (page) => {
       const res = changePage(page);
       if(res) {
-        res.then(updateData)
+        res.then(update)
         .catch(() => hasError.value = true)
         .finally(() => loading.value = false);
       }
@@ -83,37 +119,50 @@ export default defineComponent({
 
     /**
      *
-     * @param {import("@etalab/explore.data.gouv.fr/lib/csvapi").CsvapiResponse} response
+     * @param {string} col
      */
-    const updateData = (response) => {
-      if (response.ok) {
-        rows.value = response.rows;
-        columns.value = response.columns;
-        rowCount.value = response.total;
-        columnCount.value = response.columns.length;
-        totalRows.value = response.total;
-        configure({totalRows: totalRows.value});
+    const sortbyfield = (col) => {
+      if(sortBy.value == col) {
+        sortDesc.value = !sortDesc.value
       } else {
-        hasError.value = true;
+        sortDesc.value = false
       }
+      sortBy.value = col;
+      return sort(sortBy.value, sortDesc.value).then(res => {
+        update(res);
+        currentPage.value = 1;
+      }).catch(() => hasError.value = true)
+      .finally(() => loading.value = false);
+    };
+
+    const requestData = () => {
+      loading.value = true;
+      return getData("apify").then(res => {
+        update(res);
+      }).catch(() => hasError.value = true)
+      .finally(() => loading.value = false);
     }
 
-    configure({ csvapiUrl: tabular_csvapi_url, pageSize, currentPage: currentPage.value });
-    apify(props.resource.url).then(res => {
-      if (res.ok) {
-        configure({ dataEndpoint: res.endpoint });
-        return getData("apify")
-        .then(updateData)
-        .catch(() => hasError.value = true)
-        .finally(() => loading.value = false);
-      } else {
+    const requestApify = () => {
+      return apify(props.resource.url).then(res => {
+        if (res.ok) {
+          configure({ dataEndpoint: res.endpoint });
+          return requestData();
+        } else {
+          hasError.value = true;
+          loading.value = false;
+        }
+      }).catch(() => {
         hasError.value = true;
         loading.value = false;
-      }
-    }).catch(() => {
-      hasError.value = true;
-      loading.value = false;
-    });
+      });
+    }
+
+    const isSortedBy = (col) => col === sortBy.value;
+
+    configure({ csvapiUrl: tabular_csvapi_url, pageSize, currentPage: currentPage.value });
+
+    requestApify();
 
     return {
       hasError,
@@ -122,11 +171,14 @@ export default defineComponent({
       rows,
       rowCount,
       columnCount,
+      sortbyfield,
+      isSortedBy,
+      sortDesc,
       currentPage,
       pageSize,
       totalRows,
       changeExplorePage,
-    }
+    };
   }
 });
 </script>